+++ src/commands/index.ts
@@ -1,4 +1,3 @@
-import * as os from "os";
 import {
   commands,
   CommentReply,
@@ -16,13 +15,6 @@ import {
   ProgressLocation,
   TextEditorDecorationType,
   DecorationRangeBehavior,
-  WorkspaceEdit,
-  NotebookCellData,
-  NotebookCellKind,
-  NotebookRange,
-  NotebookEdit,
-  NotebookDocument,
-  NotebookEditor,
 } from "vscode";
 import { SqlPreviewContentProvider } from "../content_provider/sqlPreviewContentProvider";
 import { RunModelType } from "../domain";
@@ -53,10 +45,8 @@ import { existsSync, readFileSync } from "fs";
 import { DBTProject } from "../manifest/dbtProject";
 import { SQLLineagePanel } from "../webview_provider/sqlLineagePanel";
 import { QueryManifestService } from "../services/queryManifestService";
-import { NotebookFileSystemProvider } from "../file_system_providers/notebookFileSystemProvider";
-import { join } from "path";
 import { AltimateRequest } from "../altimate";
-import { RawNotebookCell, RawNotebook } from "../notebook_provider/types";
+import { DatapilotNotebookController } from "../notebook_provider/controller";
 
 @provideSingleton(VSCodeCommands)
 export class VSCodeCommands implements Disposable {
@@ -78,6 +68,7 @@ export class VSCodeCommands implements Disposable {
     private sqlLineagePanel: SQLLineagePanel,
     private queryManifestService: QueryManifestService,
     private altimate: AltimateRequest,
+    private notebookController: DatapilotNotebookController,
   ) {
     this.disposables.push(
       commands.registerCommand(
@@ -576,167 +567,7 @@ export class VSCodeCommands implements Disposable {
         async (
           args: { notebookId?: string; fileName?: string } | undefined,
         ) => {
-          const { notebookId, fileName } = args || {};
-          // TODO: is this needed
-          // workspace.updateWorkspaceFolders(0, 0, {
-          //   uri: Uri.parse("memfs:/"),
-          //   name: "MemFS - Sample",
-          // });
-          const project =
-            await this.queryManifestService.getOrPickProjectFromWorkspace();
-          if (!project) {
-            window.showErrorMessage("No dbt project selected.");
-            return;
-          }
-
-          const memFs = new NotebookFileSystemProvider();
-          const fileNamePrefix = notebookId || fileName || "poweruser";
-          const uri = Uri.parse(
-            `${project.projectRoot}/${fileNamePrefix}-${Date.now()}.notebook`,
-          ).with({ scheme: "untitled" });
-          //         const hashedProjectRoot = DBTProject.hashProjectRoot(project.projectRoot.fsPath);
-          // const tempFolder = join(os.tmpdir(), hashedProjectRoot);
-
-          // const uri = Uri.parse(`${tempFolder}/untitled.notebook`, true).with({ scheme: "memfs" });;
-          // await memFs.writeFile(uri, Buffer.from(''), { create: true, overwrite: true });
-          // const cell = new NotebookCellData(NotebookCellKind.Code, "select * ", "jinja-sql");
-          // cell.metadata = {"hello": "j"};
-          // const data = new NotebookData([cell]);
-          // data.metadata = {
-          //   custom: {
-          //     cells: [],
-          //     metadata: {
-          //       orig_nbformat: 4
-          //     },
-          //     nbformat: 4,
-          //     nbformat_minor: 2
-          //   }
-          // };
-          // workspace.openNotebookDocument("datapilot-notebook", data).then(
-          workspace.openNotebookDocument(uri).then(
-            (doc) => {
-              // set this to sql language so we can bind codelens and other features
-              // languages.setTextDocumentLanguage(doc, "jinja-sql");
-              window.showNotebookDocument(doc).then(
-                (editor) => {
-                  if (!notebookId) {
-                    return;
-                  }
-                  const notebooks =
-                    this.dbtProjectContainer.getFromGlobalState("notebooks") ||
-                    {};
-                  const contents = notebooks[notebookId];
-
-                  let raw: RawNotebookCell[];
-                  try {
-                    raw = (<RawNotebook>JSON.parse(contents)).cells;
-                  } catch {
-                    raw = [
-                      {
-                        cell_type: "code",
-                        source: [],
-                        languageId: "jinja-sql",
-                      },
-                    ];
-                  }
-
-                  const cellData = raw.map(
-                    (item) =>
-                      new NotebookCellData(
-                        NotebookCellKind.Code,
-                        item.source?.join("\n"),
-                        "jinja-sql",
-                      ),
-                  );
-
-                  // Get the active notebook editor
-                  const notebookEditor = window.activeNotebookEditor;
-                  if (notebookEditor) {
-                    // Create notebook cells
-                    const cells = cellData.map(
-                      (data) =>
-                        new NotebookCellData(
-                          data.kind,
-                          data.value,
-                          data.languageId,
-                        ),
-                    );
-
-                    // Function to backup the state of the notebook
-                    function backupNotebookState(notebook: NotebookDocument) {
-                      return notebook
-                        .getCells()
-                        .map((cell) => cell.document.getText());
-                    }
-
-                    // Function to restore the state of the notebook
-                    async function restoreNotebookState(
-                      notebook: NotebookDocument,
-                      backup: string[],
-                    ) {
-                      const edit = new WorkspaceEdit();
-                      notebook.getCells().forEach((cell, index) => {
-                        edit.replace(
-                          cell.document.uri,
-                          new Range(0, 0, cell.document.lineCount, 0),
-                          backup[index],
-                        );
-                      });
-                      await workspace.applyEdit(edit);
-                    }
-
-                    async function applyCellsWithoutDirty(
-                      notebookEditor: NotebookEditor,
-                      cells: NotebookCellData[],
-                    ) {
-                      const backup = backupNotebookState(
-                        notebookEditor.notebook,
-                      );
-
-                      // Apply the cell data to the notebook
-                      const edit = new WorkspaceEdit();
-                      edit.set(notebookEditor.notebook.uri, [
-                        new NotebookEdit(new NotebookRange(0, 0), cells),
-                      ]);
-
-                      await workspace.applyEdit(edit);
-
-                      // Restore the original state to make it appear as not dirty
-                      await restoreNotebookState(
-                        notebookEditor.notebook,
-                        backup,
-                      );
-                    }
-
-                    applyCellsWithoutDirty(notebookEditor, cells).then(() => {
-                      console.log(
-                        "Cells applied without marking the document as dirty.",
-                      );
-                    });
-                    // Apply the cell data to the notebook
-                    // const edit = new WorkspaceEdit();
-                    // // edit.replace(notebookEditor.notebook.uri, new NotebookEdit(new NotebookRange(0, 0), cells));
-                    // edit.set(notebookEditor.notebook.uri, [
-                    //   new NotebookEdit(new NotebookRange(0, 0), cells),
-                    // ]);
-                    // workspace.applyEdit(edit).then(() => {
-                    //   // notebookEditor.notebook.isDirty = false;
-                    //   // Save the notebook
-                    //   // setTimeout(() => {
-                    //   //   notebookEditor.notebook.save();
-                    //   // }, 100);
-                    // });
-                  }
-                },
-                (e) => {
-                  window.showErrorMessage((e as Error).message);
-                },
-              );
-            },
-            (e) => {
-              window.showErrorMessage((e as Error).message);
-            },
-          );
+          this.notebookController.createNotebook(args);
         },
       ),
       commands.registerCommand(
