+++ src/dbtPowerUserExtension.ts
@@ -22,7 +22,6 @@ import { DbtPowerUserActionsCenter } from "./quickpick";
 import { ValidationProvider } from "./validation_provider";
 import { CommentProviders } from "./comment_provider";
 import { NotebookProviders } from "./notebook_provider";
-import { FilesystemProviders } from "./file_system_providers";
 
 enum PromptAnswer {
   YES = "Yes",
